Revamped the graph-feature code for proper functioning
